+++ Program_By_Prof_Svet/Detail.h
@@ -7,15 +7,15 @@ private:String^ name;
 private:int model;
 private:String^ description;
 private:String^ id;
-private:int price;
+private:double price;
 private:int count;
-public:Detail(String^name,int model, String^ description, int price, int count);
+public:Detail(String^name,int model, String^ description, double price, int count);
 
 
 public:String^ getName();
 public:int getModel();
 public:String^ getDescription();
-public:int getPrice();
+public:double getPrice();
 public:int getCount();
 public:void decreaseCount(int count);
 public:String^ getInfo();
